プロジェクト管理画面 新規登録/編集機能 追加

+++ src/SkillManager/src/main/java/com/umayoryo/skill/manager/persistence/engineer/EngineerRepository.java
@@ -17,6 +17,6 @@ import com.umayoryo.skill.manager.persistence.engineer.entity.EngineerEntity;
 @Repository
 @Transactional
 public interface EngineerRepository
-        extends JpaRepository<EngineerEntity, Integer>, JpaSpecificationExecutor<EngineerEntity> {
+        extends JpaRepository<EngineerEntity, Long>, JpaSpecificationExecutor<EngineerEntity> {
 
 }
